chore: remove comments from notification replies handler
